% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{rgeopl-package}
\alias{rgeopl}
\alias{rgeopl-package}
\title{rgeopl: Unified Access to Polish National Geodata (GUGiK and BDL)}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Search and download spatial data from the Polish national geodetic resource (GUGiK / PZGiK) and the Forest Data Bank (BDL, Lasy Panstwowe), selected by an area of interest. Handles server-side record limits by transparent pagination, keeps archival vintages as a queryable attribute, and caches both index responses and downloaded files on disk.
}
\author{
\strong{Maintainer}: Igor Pawelec \email{gorasz44@gmail.com}

Authors:
\itemize{
  \item Igor Pawelec \email{gorasz44@gmail.com}
}

}
\keyword{internal}
